feature - Add superclass to extend
